pyproject.toml
[project]
name = "dc-py-bot-tmpl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "discord-py>=2.6.4",
    "dotenv>=0.9.9",
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "pandoc>=2.4",
    "prettyprint>=0.1.5",
    "sqlalchemy>=2.0.44",
]
